refactor(ai-agent): extract CatalogForwarder into dedicated module
