+++ crates/aionui-ai-agent/src/manager/acp/mod.rs
@@ -1,9 +1,11 @@
+pub mod catalog_forwarder;
 pub mod events;
 pub mod permission_router;
 pub mod reconcile;
 pub mod session;
 pub mod session_sync;
 
+pub use catalog_forwarder::CatalogForwarder;
 pub use events::AcpSessionEvent;
 pub use permission_router::PermissionRouter;
 pub use reconcile::ReconcileAction;
